Search Bar Dropdown can be Clicked
This was more complex than expected, since the `onBlur` function of the
input field hides the dropdown, hence a click cannot be done. A click is
a `mouseDown` and a `mouseUp` in the same field, but since the `mouseDown`
leads to hiding the entire dropdown, no full click can be done.

The solution is to use `onMouseDown` and use a `preventDefault()` call
to prevent the input from becoming defocused and hence calling `onBlur`.